refacto: move methods
